+++ packages/ibm-products/src/components/ConditionBuilder/ConditionGroupBuilder/ConditionGroupBuilder.js
@@ -1,4 +1,4 @@
-import React, { Fragment, useContext, useState } from 'react';
+import React, { Fragment, useContext, useRef, useState } from 'react';
 import ConditionBlock from '../ConditionBlock/ConditionBlock';
 import PropTypes from 'prop-types';
 
@@ -7,16 +7,15 @@ import cx from 'classnames';
 import {
   blockClass,
   statementConfig,
-  translateWithId,
 } from '../ConditionBuilderContext/DataConfigs';
 import { ConditionBuilderItem } from '../ConditionBuilderItem/ConditionBuilderItem';
-
-import { focusThisField } from '../utils/util';
+import { focusThisField, manageTabIndexAndFocus } from '../utils/util';
 import ConditionConnector from '../ConditionBuilderConnector/ConditionConnector';
 import { ConditionBuilderContext } from '../ConditionBuilderContext/ConditionBuilderProvider';
 import uuidv4 from '../../../global/js/utils/uuidv4';
 import ConditionPreview from '../ConditionPreview/ConditionPreview';
 import { ItemOption } from '../ConditionBuilderItem/ConditionBuilderItemOption/ItemOption';
+import { useTranslations } from '../utils/useTranslations';
 /**
  *
  *  state - this is the current group that is being rendered . This can be a inner group or outer group
@@ -29,22 +28,67 @@ const ConditionGroupBuilder = ({
   aria,
   onRemove,
   onChange,
-  conditionBuilderRef,
   className,
 }) => {
-  const { variant } = useContext(ConditionBuilderContext);
+  const [conditionBuilderGroupText, conditionText, conditionBuilderText] =
+    useTranslations([
+      'conditionBuilderGroupText',
+      'conditionText',
+      'conditionBuilderText',
+    ]);
+  const { variant, conditionBuilderRef } = useContext(ConditionBuilderContext);
   const [showConditionPreview, setShowConditionPreview] = useState(-1);
   const [showConditionSubGroupPreview, setShowConditionSubGroupPreview] =
     useState(-1);
-  const onRemoveHandler = (conditionId, evt) => {
+  const conditionBuilderContentRef = useRef();
+  const onRemoveHandler = (conditionId, evt, conditionIndex) => {
     if (group.conditions.length > 1) {
-      onChange({
-        ...group,
-        conditions: group.conditions.filter(
-          (condition) => conditionId !== condition.id
-        ),
-      });
-      handleFocusOnClose(evt);
+      variant == 'tree'
+        ? handleFocusOnCloseTree(evt)
+        : handleFocusOnClose(evt, conditionIndex);
+
+      if (
+        group.conditions[1].conditions &&
+        group.conditions[1].id !== conditionId
+      ) {
+        //when we remove every plain conditions of a group without deleting  the subgroup, we need to restructure the group
+        //the inner group become outer group and same level subgroups become plain conditions
+
+        //ensure we are deleting last condition , not the subgroup
+
+        //spreading out the condition inside the subgroup
+        const allConditions = group.conditions.reduce((acc, item) => {
+          if (item.conditions) {
+            return acc.concat(item.conditions);
+          }
+          return acc;
+        }, []);
+
+        //we always have conditions first and then subgroups, so ordering accordingly
+        const groupedItems = {
+          groups: [],
+          conditions: [],
+        };
+        allConditions.forEach((item) => {
+          if (item.conditions) {
+            groupedItems.groups.push(item);
+          } else {
+            groupedItems.conditions.push(item);
+          }
+        });
+
+        onChange({
+          ...group,
+          conditions: [...groupedItems.conditions, ...groupedItems.groups],
+        });
+      } else {
+        onChange({
+          ...group,
+          conditions: group.conditions.filter(
+            (condition) => conditionId !== condition.id
+          ),
+        });
+      }
     } else {
       onRemove(evt);
     }
@@ -81,13 +125,89 @@ const ConditionGroupBuilder = ({
     });
   };
 
-  const handleFocusOnClose = (e) => {
-    const previousClose = e.currentTarget
+  const handleFocusOnClose = (e, conditionIndex) => {
+    //get all close buttons.
+    //if the last condition is closing, focus the second last one.
+    //or focus the next one.
+    const currentGroupCloseButtons = e.currentTarget
+      .closest(`.${blockClass}__group`)
+      ?.querySelectorAll('[data-name="closeCondition"]');
+    if (conditionIndex == currentGroupCloseButtons.length - 1) {
+      manageTabIndexAndFocus(
+        currentGroupCloseButtons[conditionIndex - 1],
+        conditionBuilderRef
+      );
+    } else {
+      manageTabIndexAndFocus(
+        currentGroupCloseButtons[conditionIndex + 1],
+        conditionBuilderRef
+      );
+    }
+  };
+  const handleFocusOnCloseTree = (evt) => {
+    //getting the current aria-level and aria-posinset.
+    const currentLevel = evt.currentTarget
       ?.closest('[role="row"]')
-      ?.previousSibling?.querySelector('[data-name="closeCondition"]');
+      ?.getAttribute('aria-level');
+    const currentPos = evt.currentTarget
+      ?.closest('[role="row"]')
+      ?.getAttribute('aria-posinset');
 
-    if (previousClose) {
-      previousClose.focus();
+    //finding the next and previous items in same level
+    const nextElement = conditionBuilderContentRef.current?.querySelector(
+      `[aria-level="${currentLevel}"][aria-posinset="${
+        Number(currentPos) + 1
+      }"]`
+    );
+    const prevElement = conditionBuilderContentRef.current?.querySelector(
+      `[aria-level="${currentLevel}"][aria-posinset="${
+        Number(currentPos) - 1
+      }"]`
+    );
+    //checking if next level is a valid condition. If then, focus the close button inside that condition
+    //Otherwise , check the previous item is a valid condition
+
+    if (nextElement?.classList.contains(`${blockClass}__condition-block`)) {
+      manageTabIndexAndFocus(
+        nextElement?.querySelector('[data-name="closeCondition"]'),
+        conditionBuilderRef
+      );
+    } else if (
+      prevElement?.classList.contains(`${blockClass}__condition-block`)
+    ) {
+      manageTabIndexAndFocus(
+        prevElement?.querySelector('[data-name="closeCondition"]'),
+        conditionBuilderRef
+      );
+    }
+    //If there are no valid condition in this group, focus next or previous row
+    else {
+      const prevRows = conditionBuilderContentRef.current?.querySelectorAll(
+        `[aria-level="${Number(currentLevel) - 1}"][role="row"]`
+      );
+      const nextRow = conditionBuilderContentRef.current?.querySelector(
+        `[aria-level="${Number(currentLevel) + 1}"][role="row"]`
+      );
+      if (nextRow) {
+        //since there are no condition in current group, this group will move one level up
+
+        const rowIdentity = {
+          ariaLevel: Number(nextRow.ariaLevel) - 1,
+          ariaPosInSet: nextRow.ariaPosInSet,
+        };
+        setTimeout(() => {
+          const currentRowToFocus =
+            conditionBuilderContentRef.current.querySelector(
+              `[role="row"][aria-level="${rowIdentity.ariaLevel}"][aria-posinset="${rowIdentity.ariaPosInSet}"]`
+            );
+          manageTabIndexAndFocus(currentRowToFocus, conditionBuilderRef);
+        }, 0);
+      } else if (prevRows?.length > 1) {
+        manageTabIndexAndFocus(
+          prevRows[prevRows.length - 2],
+          conditionBuilderRef
+        );
+      }
     }
   };
 
@@ -135,8 +255,12 @@ const ConditionGroupBuilder = ({
     setShowConditionPreview(-1);
   };
   const onStatementChangeHandler = (updatedStatement) => {
+    const groupOperator = statementConfig.find(
+      (statement) => statement.id == updatedStatement
+    )?.connector;
     onChange({
       ...group,
+      groupOperator: groupOperator,
       statement: updatedStatement,
     });
   };
@@ -154,7 +278,7 @@ const ConditionGroupBuilder = ({
         <div
           className={`${blockClass}__condition-wrapper`}
           role="grid"
-          aria-label={translateWithId('condition_builder_group')}
+          aria-label={conditionBuilderText}
         >
           {/* condition loop starts here */}
 
@@ -181,7 +305,7 @@ const ConditionGroupBuilder = ({
                 }}
                 addConditionHandler={addConditionHandler}
                 onRemove={(e) => {
-                  onRemoveHandler(eachCondition.id, e);
+                  onRemoveHandler(eachCondition.id, e, conditionIndex);
                 }}
                 onConnectorOperatorChange={onConnectorOperatorChange}
                 onStatementChange={onStatementChangeHandler}
@@ -199,11 +323,8 @@ const ConditionGroupBuilder = ({
       <div
         className={`${className} ${blockClass}__condition-wrapper`}
         role={aria.level === 1 ? 'rowgroup' : undefined}
-        aria-label={
-          aria.level == 1
-            ? translateWithId('condition_builder_group')
-            : undefined
-        }
+        aria-label={aria.level == 1 ? conditionBuilderGroupText : undefined}
+        ref={conditionBuilderContentRef}
       >
         {/* condition loop starts here */}
         {
@@ -216,7 +337,7 @@ const ConditionGroupBuilder = ({
           >
             <ConditionBuilderItem
               label={group.statement}
-              title={translateWithId('condition')}
+              title={conditionText}
               data-name="connectorField"
               popOverClassName={`${blockClass}__gap`}
               className={`${blockClass}__statement-button`}
@@ -224,14 +345,14 @@ const ConditionGroupBuilder = ({
               <ItemOption
                 conditionState={{
                   value: group.statement,
-                  label: translateWithId('condition'),
+                  label: conditionText,
                 }}
                 onChange={(v, evt) => {
-                  focusThisField(evt);
+                  focusThisField(evt, conditionBuilderRef);
                   onStatementChangeHandler(v);
                 }}
                 config={{ options: statementConfig }}
-              ></ItemOption>
+              />
             </ConditionBuilderItem>
           </div>
         }
@@ -251,7 +372,7 @@ const ConditionGroupBuilder = ({
                 )}
               >
                 <ConditionConnector
-                  className={`${blockClass}__gap ${blockClass}__groupConnector`}
+                  className={`${blockClass}__gap ${blockClass}__gap-bottom ${blockClass}__groupConnector`}
                   operator={group.groupOperator}
                   aria-hidden
                 />
@@ -270,7 +391,6 @@ const ConditionGroupBuilder = ({
                   onRemove={(e) => {
                     onRemoveHandler(eachCondition.id, e);
                   }}
-                  conditionBuilderRef={conditionBuilderRef}
                 />
               </div>
             ) : (
@@ -314,10 +434,10 @@ const ConditionGroupBuilder = ({
             )}
 
             {conditionIndex == showConditionSubGroupPreview && (
-              <ConditionPreview previewType="subGroup" />
+              <ConditionPreview previewType="subGroup" group={group} />
             )}
             {conditionIndex == showConditionPreview && (
-              <ConditionPreview previewType="condition" />
+              <ConditionPreview previewType="condition" group={group} />
             )}
           </Fragment>
         ))}
@@ -347,11 +467,6 @@ ConditionGroupBuilder.propTypes = {
    */
   className: PropTypes.string,
 
-  /**
-  className: PropTypes.string,
-   * ref of condition builder
-   */
-  conditionBuilderRef: PropTypes.object,
   group: PropTypes.object,
   /**
    * callback to update the current condition of the state tree