+++ packages/ibm-products/src/components/Datagrid/types/index.ts
@@ -26,11 +26,13 @@ import {
   TableCommonProps,
   TableDispatch,
   TableInstance,
+  TableState,
   TableToggleAllRowsSelectedProps,
   UseExpandedRowProps,
   UseFiltersInstanceProps,
   UsePaginationInstanceProps,
   UseResizeColumnsColumnProps,
+  UseResizeColumnsOptions,
   UseResizeColumnsState,
   UseRowSelectInstanceProps,
   UseRowSelectRowProps,
@@ -149,7 +151,9 @@ export interface DatagridTableHooks<T extends object = any>
 
 export interface DatagridColumn<T extends object = any>
   extends ColumnInstance<T>,
-    UseSortByOptions<T> {
+    UseSortByOptions<T>,
+    Partial<UseResizeColumnsColumnProps<T>>,
+    UseResizeColumnsOptions<T> {
   sticky?: 'left' | 'right';
   className?: string;
   rightAlignedColumn?: boolean;
@@ -187,11 +191,7 @@ export interface DataGridHeaderGroup<T extends object = any>
   extends HeaderGroup<T>,
     UseResizeColumnsColumnProps<T> {}
 
-export interface TableProps {
-  className?: string;
-  role?: string;
-  style?: CSSStyleDeclaration;
-}
+export interface DataGridTableProps extends TableCommonProps {}
 
 interface DataGridTableState
   extends UseResizeColumnsState<any>,
@@ -200,7 +200,18 @@ interface DataGridTableState
 }
 
 export interface DataGridTableInstance<T extends object = any>
-  extends TableInstance<T> {}
+  extends Omit<TableInstance<T>, 'state'>,
+    Partial<UsePaginationInstanceProps<any>> {
+  shouldDisableSelectRow?: (...args) => void | boolean;
+  state?: Partial<TableState & UseRowSelectState<any>>;
+  disableSelectAll?: boolean;
+  disableSelectRowsProps?: {
+    labels?: {
+      toggleAllRowsLabel?: string;
+    };
+  };
+  withSelectRows?: boolean;
+}
 
 export interface RowAction {
   id?: string;
@@ -316,6 +327,12 @@ export interface DataGridState<T extends object = any>
   ExpandedRowContentComponent?: JSXElementConstructor<any>;
 }
 
+export interface DataGridData {
+  instance?: DataGridTableInstance;
+  column?: DatagridColumn;
+  cell?: DataGridCell;
+}
+
 // DatagridHeaderRow related types
 export interface ResizeHeaderProps {
   resizerProps?: ResizerProps;
@@ -339,3 +356,8 @@ export type VisibleColumns<T extends object = {}> = (
 ) => Array<Column<T>>;
 
 export type NodeFuncType = (props) => ReactNode;
+
+export interface PropGetterMeta {
+  instance?: DataGridTableInstance;
+  row?: Partial<Row<any> & DatagridRow<any>>;
+}