+++ packages/ibm-products/src/components/SidePanel/SidePanel.tsx
@@ -16,7 +16,7 @@ import React, {
   MutableRefObject,
   RefObject,
 } from 'react';
-import { motion, AnimatePresence, useReducedMotion } from 'framer-motion';
+import { motion, AnimatePresence } from 'framer-motion';
 
 // Other standard imports.
 import PropTypes from 'prop-types';
@@ -41,6 +41,7 @@ import {
 } from './motion/variants';
 import pconsole from '../../global/js/utils/pconsole';
 import { ButtonProps } from '@carbon/react';
+import usePrefersReducedMotion from '../../global/js/hooks/usePrefersReducedMotion';
 
 const blockClass = `${pkg.prefix}--side-panel`;
 const componentName = 'SidePanel';
@@ -284,7 +285,7 @@ export let SidePanel = React.forwardRef(
       .current;
     const previousOpen = usePreviousValue(open);
 
-    const shouldReduceMotion = useReducedMotion();
+    const shouldReduceMotion = usePrefersReducedMotion();
 
     // Title animation on scroll related state
     const [labelTextHeight, setLabelTextHeight] = useState<any>(0);
@@ -350,11 +351,6 @@ export let SidePanel = React.forwardRef(
       titleItemsStyles,
     ]);
 
-    const reducedMotion =
-      typeof document !== 'undefined' && window?.matchMedia
-        ? window.matchMedia('(prefers-reduced-motion: reduce)')
-        : { matches: true };
-
     // scroll panel to top going between steps
     useEffect(() => {
       if (sidePanelRef && panelRefValue) {
@@ -410,7 +406,7 @@ export let SidePanel = React.forwardRef(
         titleRef?.current &&
         title &&
         title.length &&
-        !reducedMotion.matches
+        !shouldReduceMotion
       ) {
         const titleEl = titleRef.current;
         const labelHeight = labelTextRef?.current?.offsetHeight ?? 0;
@@ -485,7 +481,7 @@ export let SidePanel = React.forwardRef(
         titleRef?.current &&
         title &&
         title.length &&
-        !reducedMotion.matches
+        !shouldReduceMotion
       ) {
         checkSetDoAnimateTitle();
       }
@@ -498,7 +494,7 @@ export let SidePanel = React.forwardRef(
       handleScroll,
       title,
       size,
-      reducedMotion.matches,
+      shouldReduceMotion,
       id,
     ]);
 
@@ -549,10 +545,10 @@ export let SidePanel = React.forwardRef(
     // Set the internal state `animationComplete` to true if
     // prefers reduced motion is true
     useEffect(() => {
-      if (reducedMotion.matches) {
+      if (shouldReduceMotion) {
         setAnimationComplete(true);
       }
-    }, [reducedMotion.matches]);
+    }, [shouldReduceMotion]);
 
     // initializes the side panel to open
     const onAnimationStart = () => {
@@ -578,11 +574,11 @@ export let SidePanel = React.forwardRef(
         !open &&
         previousState &&
         previousState['open'] &&
-        reducedMotion.matches
+        shouldReduceMotion
       ) {
         onUnmount?.();
       }
-    }, [open, onUnmount, reducedMotion.matches, previousState]);
+    }, [open, onUnmount, shouldReduceMotion, previousState]);
 
     // used to set margins of content for slide in panel version
     useEffect(() => {
@@ -599,13 +595,13 @@ export let SidePanel = React.forwardRef(
         }
         if (placement && placement === 'right' && pageContentElement) {
           pageContentElement.style.marginInlineEnd = '0';
-          pageContentElement.style.transition = !reducedMotion.matches
+          pageContentElement.style.transition = !shouldReduceMotion
             ? `margin-inline-end ${moderate02}`
             : '';
           pageContentElement.style.marginInlineEnd = SIDE_PANEL_SIZES[size];
         } else if (pageContentElement) {
           pageContentElement.style.marginInlineStart = '0';
-          pageContentElement.style.transition = !reducedMotion.matches
+          pageContentElement.style.transition = !shouldReduceMotion
             ? `margin-inline-start ${moderate02}`
             : '';
           pageContentElement.style.marginInlineStart = SIDE_PANEL_SIZES[size];
@@ -616,7 +612,7 @@ export let SidePanel = React.forwardRef(
       selectorPageContent,
       placement,
       size,
-      reducedMotion.matches,
+      shouldReduceMotion,
       open,
     ]);
 
@@ -672,7 +668,7 @@ export let SidePanel = React.forwardRef(
           {title}
         </h2>
 
-        {doAnimateTitle && !reducedMotion.matches && (
+        {doAnimateTitle && !shouldReduceMotion && (
           <h2
             className={`${blockClass}__collapsed-title-text`}
             title={title}
@@ -700,7 +696,7 @@ export let SidePanel = React.forwardRef(
           className={cx(`${blockClass}__header`, {
             [`${blockClass}__header--on-detail-step`]: currentStep > 0,
             [`${blockClass}__header--no-title-animation`]: !animateTitle,
-            [`${blockClass}__header--reduced-motion`]: reducedMotion.matches,
+            [`${blockClass}__header--reduced-motion`]: shouldReduceMotion,
             [`${blockClass}__header--has-title`]: title,
           })}
           ref={headerRef}
@@ -708,7 +704,6 @@ export let SidePanel = React.forwardRef(
           {/* back button */}
           {currentStep > 0 && (
             <Button
-              aria-label={navigationBackIconDescription}
               kind="ghost"
               size={slugCloseSize}
               disabled={false}